🔥 GameEngine: Users on room paradigm
Removed Google authentication and using an ephemeral user. Chosen to allow for easy to hop in and hop off. Since this isn't a serious platform to begin with, this seems the more logical choice.

If in the future would want to include user stats, could return with user auth and simply tie stats to users.